requirements.txt
av==10.0.0
chumpy == 0.70
imageio-ffmpeg==0.4.8
jsonargparse[all] == 4.22.1
numpy == 1.24.4
open3d == 0.16
Pillow == 10.0.0
pyembree == 0.2.11
pyglet < 2
pytorch-lightning == 1.7.7
pyvista == 0.40.1
smplx[all] == 0.1.28
torch == 1.13.1
torchmetrics == 0.10.0
